// apps/web/lib/types.ts
export interface UserProps {
  id: string;
  name: string;
  email: string;
  image?: string;
  createdAt: Date;
  role: "OWNER" | "MEMBER";
  workspaces?: { workspaceId: string }[];
  sites?: { domain: string; key: string; url: string }[];
}